Get a function which returns a table object

Make a function that returns a table getter. The function which is returned can be called to produce a table_node object.

Usage

make_table_getter(con, database, table_schema, table_name, id = NULL)

Arguments

con

The database connection (from server)

database

The database name

table_schema

The table schema name

table_name

The table name

id

You can also pass the complete Id, if it is available

Details

This function is the way to associate a function with every table object in the db list, and replaces the need to overload $, while also avoiding the problem of storing the table name and database name in the functions environment. This is the purpose of a function factory; to capture variables in the enclosing environment and allow them to persist when the function is called. Read this page and the associated environment sections for a full explanation: https://adv-r.hadley.nz/function-factories.html

Even though the code uses () to access the table name, this level of indirection is still necessary because of the bug referenced in the build_object_tree body -- using a table getter means that the tbl is only created when the user asks for it, removing some edge cases with "bad" tables (like COLUMNS_EXTENSIONS in mysql information_schema, which contains JSON columns).