Mapping-tree parser

This function parses the tree returned by reading the yaml mapping file, and returns a named list of the contents of the current level passed as the argument. The function is recursive, and will descend through the configuration, generating a tree of documentation, database and table information that acts like a logical database.

Usage

parse_mapping(mapping, srv, source_database = NULL, source_table = NULL)

Arguments

mapping

A named list storing a level of the yaml config file

srv

The Databases object with the underlying connection

source_database

The name of the source database for the current level

source_table

The name of the source table for the current level

Value

A named list containing the results of parsing the file